requirements.txt
configparser 
python-telegram-bot==22.5 
urllib3==2.6.2 
urllib3==2.6.2 
